moved delta messages to separate namespace
